adding the steam turbine component to the component module init
